--- app/(authentication)/profile/update/page.jsx
@@ -1,7 +1,0 @@
-export default function Profile() {
-  return (
-    <div className='flex'>
-      <h>Profile</h>
-    </div>
-  );
-}
